+++ runner.py
@@ -9,7 +9,7 @@ Keep in mind, that not all logs will be shown in bot_runner's terminal.
 """
 
 if __name__ == "__main__":
-    cmd = "python3.6 -u bot.py"
+    cmd = "python3.6 -u run.py"
     p = None
 
     try: